+++ src/index.ts
@@ -1,9 +1,11 @@
+import { BroadcastsRecorder } from './broadcasts-recorder';
 import { Context } from './context';
 import { EventManager } from './manager';
 import { Pipeline } from './pipeline';
 import { Procedure } from './procedure';
 
 export {
+  BroadcastsRecorder,
   Context,
   EventManager,
   Pipeline,